% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_range_generate_labels.R
\name{generate_gift_range_labels}
\alias{generate_gift_range_labels}
\title{Helper function to dynamically generate labels (Internal)}
\usage{
generate_gift_range_labels(upper_limits, lower_bound, prefixes = NULL)
}
\description{
Helper function to dynamically generate labels (Internal)
}
\keyword{internal}
